+++ src/librt/primitives/poly/poly.c
@@ -63,7 +63,7 @@ rt_pg_prep(struct soltab *stp, struct rt_db_internal *ip, struct rt_i *rtip)
 {
     struct rt_pg_internal *pgp;
     int i;
-    int p;
+    ssize_t p;
 
     pgp = (struct rt_pg_internal *)ip->idb_ptr;
     RT_PG_CK_MAGIC(pgp);
@@ -814,7 +814,7 @@ rt_pg_describe(struct bu_vls *str, const struct rt_db_internal *ip, int verbose,
     RT_PG_CK_MAGIC(pgp);
     bu_vls_strcat(str, "polygon solid with no topology (POLY)\n");
 
-    sprintf(buf, "\t%d polygons (faces)\n",
+    sprintf(buf, "\t%ld polygons (faces)\n",
 	    pgp->npoly);
     bu_vls_strcat(str, buf);
